+++ src/grid/impls.rs
@@ -5,6 +5,8 @@ use crate::{
     grid::{GridReadUnchecked, GridWriteUnchecked},
 };
 
+
+
 /// Implementation of [`GridRead::get`](crate::grid::GridRead::get) for a [`GridReadUnchecked`].
 ///
 /// The size of the grid is referenced from the grid itself.